abstract revealer from tooltip and hide tooltip on escape
