+++ app/src/Components/Revealer/Revealer.constants.ts
